+++ frontend/CHANGELOG.md
@@ -1,10 +1,15 @@
 # Changelog
 
 <a name="0.1.0"></a>
-## 0.1.0 (2021-08-14)
+## 0.1.0 (2021-08-15)
 
 ### Added
 
+- ✨ 🐛🔥Meals CRUD [548961e]
+- ✨ 💄UI Fixes [29f4209]
+- ✨ 🐛⚡Restaurant CRUD [cfe0226]
+- ✨ 🎨🐛 Block List and Cart Total [161b989]
+- ✨ 🔥🐛⚡🎨 Orders Page With Cancelation and Pagination [fb3a7f2]
 - ➕ 💄 Added Chakra UI [ad2aaa6]
 - ✨ 🎉 Frontend Folder Created with Create React App [bdd1e94]
 - ➕ Add Prisma Client [c7da42b]
@@ -13,6 +18,11 @@
 
 ### Changed
 
+- ⚡ 🐛✨ Registeration Type and Bug Fixes [962e9ea]
+- ⚡ ✨🐛Restaurant Order Page &amp; Bug Fixes [92019b0]
+- 💄 UI Updates For Loading Spinner on Orders Page [0622a8b]
+- 💄 🎨🔥🐛✨ Orders Page: New Styles and Polling [c1ff32d]
+- ⚡ 🐛✨Restaurant and Single Cart Logic [20dd3a4]
 - 🎨 🐛Fixes Token Sustain Bug and Improves Structure [7efee7b]
 - 💄 🎨🐛⚡🚑 Hotfix and Protected Routes + Designs [0e5d055]
 - ⚡ 🔥🐛 Refactor GraphQL Logic + Bug Fixes [8c46ee5]
@@ -21,8 +31,14 @@
 - 🗃️ Prisma Migrations [37c6ac4]
 - 🔧 Added Typescript Configuration [8d40625]
 
+### Removed
+
+- 🔥 Remove Test Route [581249e]
+- 🔥 ⚡🐛✨Added Order Placement [b21fdc3]
+
 ### Fixed
 
+- 🐛 ✨ Single Restaurant Pages [3496edf]
 - 🚑 HotFix Login Page [cfb592e]
 - ✏️ Fixed: Array Typos in Entities [ccc966a]
 
@@ -32,6 +48,9 @@
 
 ### Miscellaneous
 
+- 🛂 🐛Authorization And Bug Fixes [61f3c97]
+- 📝 Updated READMEs [4348b81]
+- 📝 Added Changelogs [9ed3118]
 - 🛂 Authentication Complete [bd7857e]
 - 🔨 GraphQL Added With Support For CodeGen [97b61b5]
 - ⚰️ Removed Unused Imports [2ed703b]